add json implementation in foundation
